// decoupled-site/nextjs-sample/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
const defaultOptions = {} as const;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Bool: any;
  Date: any;
};

export type ArtistContainerPage = IContent & {
  __typename?: 'ArtistContainerPage';
  Ancestors?: Maybe<Array<Maybe<Scalars['String']>>>;
  Category?: Maybe<Array<Maybe<CategoryModel>>>;
  Changed?: Maybe<Scalars['Date']>;
  ContentLink?: Maybe<ContentModelReference>;
  ContentType?: Maybe<Array<Maybe<Scalars['String']>>>;
  Created?: Maybe<Scalars['Date']>;
  ExistingLanguages?: Maybe<Array<Maybe<ContentLanguageModel>>>;
  IsCommonDraft?: Maybe<Scalars['Bool']>;
  Language?: Maybe<ContentLanguageModel>;
  MasterLanguage?: Maybe<ContentLanguageModel>;
  Name?: Maybe<Scalars['String']>;
  ParentLink?: Maybe<ContentModelReference>;
  RelativePath?: Maybe<Scalars['String']>;
  RouteSegment?: Maybe<Scalars['String']>;
  Saved?: Maybe<Scalars['Date']>;
  StartPublish?: Maybe<Scalars['Date']>;
  Status?: Maybe<Scalars['String']>;
  StopPublish?: Maybe<Scalars['Date']>;
  Url?: Maybe<Scalars['String']>;
  _children?: Maybe<QueryRef>;
  _fulltext?: Maybe<Array<Maybe<Scalars['String']>>>;
  _score?: Maybe<Scalars['Float']>;
};
// ...
export type ArtistDetailsPage = IContent & {
  __typename?: 'ArtistDetailsPage';
  Ancestors?: Maybe<Array<Maybe<Scalars['String']>>>;
  /** Description to appear on the artist detail page. */
  ArtistDescription?: Maybe<Scalars['String']>;
  ArtistGenre?: Maybe<Scalars['String']>;
  ArtistIsHeadliner?: Maybe<Scalars['Bool']>;
  ArtistName?: Maybe<Scalars['String']>;
  ArtistPhoto?: Maybe<Scalars['String']>;
  Category?: Maybe<Array<Maybe<CategoryModel>>>;
  Changed?: Maybe<Scalars['Date']>;
  ContentLink?: Maybe<ContentModelReference>;
  ContentType?: Maybe<Array<Maybe<Scalars['String']>>>;
  Created?: Maybe<Scalars['Date']>;
  ExistingLanguages?: Maybe<Array<Maybe<ContentLanguageModel>>>;
  IsCommonDraft?: Maybe<Scalars['Bool']>;
  Language?: Maybe<ContentLanguageModel>;
  MasterLanguage?: Maybe<ContentLanguageModel>;
  Name?: Maybe<Scalars['String']>;
  ParentLink?: Maybe<ContentModelReference>;
  PerformanceEndTime?: Maybe<Scalars['Date']>;
  PerformanceStartTime?: Maybe<Scalars['Date']>;
  RelativePath?: Maybe<Scalars['String']>;
  RouteSegment?: Maybe<Scalars['String']>;
  Saved?: Maybe<Scalars['Date']>;
  StageName?: Maybe<Scalars['String']>;
  StartPublish?: Maybe<Scalars['Date']>;
  Status?: Maybe<Scalars['String']>;
  StopPublish?: Maybe<Scalars['Date']>;
  Url?: Maybe<Scalars['String']>;
  _children?: Maybe<QueryRef>;
  _fulltext?: Maybe<Array<Maybe<Scalars['String']>>>;
  _score?: Maybe<Scalars['Float']>;
};
// ...
export type CategoryModel = {
  __typename?: 'CategoryModel';
  Description?: Maybe<Scalars['String']>;
  Id?: Maybe<Scalars['Int']>;
  Name?: Maybe<Scalars['String']>;
};
// ...
export type Content = IContent & {
  __typename?: 'Content';
  Ancestors?: Maybe<Array<Maybe<Scalars['String']>>>;
  Changed?: Maybe<Scalars['Date']>;
  ContentLink?: Maybe<ContentModelReference>;
  ContentType?: Maybe<Array<Maybe<Scalars['String']>>>;
  Created?: Maybe<Scalars['Date']>;
  ExistingLanguages?: Maybe<Array<Maybe<ContentLanguageModel>>>;
  IsCommonDraft?: Maybe<Scalars['Bool']>;
  Language?: Maybe<ContentLanguageModel>;
  MasterLanguage?: Maybe<ContentLanguageModel>;
  Name?: Maybe<Scalars['String']>;
  ParentLink?: Maybe<ContentModelReference>;
  RelativePath?: Maybe<Scalars['String']>;
  RouteSegment?: Maybe<Scalars['String']>;
  Saved?: Maybe<Scalars['Date']>;
  StartPublish?: Maybe<Scalars['Date']>;
  Status?: Maybe<Scalars['String']>;
  StopPublish?: Maybe<Scalars['Date']>;
  Url?: Maybe<Scalars['String']>;
  _children?: Maybe<QueryRef>;
  _fulltext?: Maybe<Array<Maybe<Scalars['String']>>>;
  _score?: Maybe<Scalars['Float']>;
};
// ...
export type ContentBlock = IContent & {
  __typename?: 'ContentBlock';
  Ancestors?: Maybe<Array<Maybe<Scalars['String']>>>;
  Category?: Maybe<Array<Maybe<CategoryModel>>>;
  Changed?: Maybe<Scalars['Date']>;
  Content?: Maybe<Scalars['String']>;
  ContentLink?: Maybe<ContentModelReference>;
  ContentType?: Maybe<Array<Maybe<Scalars['String']>>>;
  Created?: Maybe<Scalars['Date']>;
  ExistingLanguages?: Maybe<Array<Maybe<ContentLanguageModel>>>;
  Image?: Maybe<Scalars['String']>;
  ImageAlignment?: Maybe<Scalars['String']>;
  IsCommonDraft?: Maybe<Scalars['Bool']>;
  Language?: Maybe<ContentLanguageModel>;
  MasterLanguage?: Maybe<ContentLanguageModel>;
  Name?: Maybe<Scalars['String']>;
  ParentLink?: Maybe<ContentModelReference>;
  RelativePath?: Maybe<Scalars['String']>;
  RouteSegment?: Maybe<Scalars['String']>;
  Saved?: Maybe<Scalars['Date']>;
  StartPublish?: Maybe<Scalars['Date']>;
  Status?: Maybe<Scalars['String']>;
  StopPublish?: Maybe<Scalars['Date']>;
  Title?: Maybe<Scalars['String']>;
  Url?: Maybe<Scalars['String']>;
  _children?: Maybe<QueryRef>;
  _fulltext?: Maybe<Array<Maybe<Scalars['String']>>>;
  _score?: Maybe<Scalars['Float']>;
};
// ...
export type ContentLanguageModel = {
  __typename?: 'ContentLanguageModel';
  DisplayName?: Maybe<Scalars['String']>;
  Link?: Maybe<Scalars['String']>;
  Name?: Maybe<Scalars['String']>;
};
// ...
export type ContentModelReference = {
  __typename?: 'ContentModelReference';
  Expanded?: Maybe<IContent>;
  GuidValue?: Maybe<Scalars['String']>;
  Id?: Maybe<Scalars['Int']>;
  Language?: Maybe<ContentLanguageModel>;
  ProviderName?: Maybe<Scalars['String']>;
  Url?: Maybe<Scalars['String']>;
  WorkId?: Maybe<Scalars['Int']>;
};
// ...
export type IContent = {
  Ancestors?: Maybe<Array<Maybe<Scalars['String']>>>;
  Changed?: Maybe<Scalars['Date']>;
  ContentLink?: Maybe<ContentModelReference>;
  ContentType?: Maybe<Array<Maybe<Scalars['String']>>>;
  Created?: Maybe<Scalars['Date']>;
  ExistingLanguages?: Maybe<Array<Maybe<ContentLanguageModel>>>;
  IsCommonDraft?: Maybe<Scalars['Bool']>;
  Language?: Maybe<ContentLanguageModel>;
  MasterLanguage?: Maybe<ContentLanguageModel>;
  Name?: Maybe<Scalars['String']>;
  ParentLink?: Maybe<ContentModelReference>;
  RelativePath?: Maybe<Scalars['String']>;
  RouteSegment?: Maybe<Scalars['String']>;
  Saved?: Maybe<Scalars['Date']>;
  StartPublish?: Maybe<Scalars['Date']>;
  Status?: Maybe<Scalars['String']>;
  StopPublish?: Maybe<Scalars['Date']>;
  Url?: Maybe<Scalars['String']>;
  _children?: Maybe<QueryRef>;
  _fulltext?: Maybe<Array<Maybe<Scalars['String']>>>;
  _score?: Maybe<Scalars['Float']>;
};
// ...
export type LandingPage = IContent & {
  __typename?: 'LandingPage';
  Ancestors?: Maybe<Array<Maybe<Scalars['String']>>>;
  ArtistsLink?: Maybe<ContentModelReference>;
  BuyTicketBlock?: Maybe<LandingPageBlockData>;
  Category?: Maybe<Array<Maybe<CategoryModel>>>;
  Changed?: Maybe<Scalars['Date']>;
  ContentLink?: Maybe<ContentModelReference>;
  ContentType?: Maybe<Array<Maybe<Scalars['String']>>>;
  Created?: Maybe<Scalars['Date']>;
  ExistingLanguages?: Maybe<Array<Maybe<ContentLanguageModel>>>;
  FooterContentArea?: Maybe<Array<Maybe<ContentAreaItemModel>>>;
  HeroImage?: Maybe<Scalars['String']>;
  IsCommonDraft?: Maybe<Scalars['Bool']>;
  Language?: Maybe<ContentLanguageModel>;
  MainContentArea?: Maybe<Array<Maybe<ContentAreaItemModel>>>;
  MasterLanguage?: Maybe<ContentLanguageModel>;
  Name?: Maybe<Scalars['String']>;
  ParentLink?: Maybe<ContentModelReference>;
  RelativePath?: Maybe<Scalars['String']>;
  RouteSegment?: Maybe<Scalars['String']>;
  Saved?: Maybe<Scalars['Date']>;
  StartPublish?: Maybe<Scalars['Date']>;
  Status?: Maybe<Scalars['String']>;
  StopPublish?: Maybe<Scalars['Date']>;
  Subtitle?: Maybe<Scalars['String']>;
  Title?: Maybe<Scalars['String']>;
  Url?: Maybe<Scalars['String']>;
  _children?: Maybe<QueryRef>;
  _fulltext?: Maybe<Array<Maybe<Scalars['String']>>>;
  _score?: Maybe<Scalars['Float']>;
};

export type LandingPageBlockData = {
  __typename?: 'LandingPageBlockData';
  Heading?: Maybe<Scalars['String']>;
  Message?: Maybe<Scalars['String']>;
};
// ...
export enum Locales {
  En = 'en',
  Sv = 'sv'
}
// ...
export type QueryRef = {
  __typename?: 'QueryRef';
  ArtistContainerPage?: Maybe<ArtistContainerPageOutput>;
  ArtistDetailsPage?: Maybe<ArtistDetailsPageOutput>;
  BuyTicketBlock?: Maybe<BuyTicketBlockOutput>;
  Content?: Maybe<ContentOutput>;
  ContentBlock?: Maybe<ContentBlockOutput>;
  ImageFile?: Maybe<ImageFileOutput>;
  ImagePage?: Maybe<ImagePageOutput>;
  LandingPage?: Maybe<LandingPageOutput>;
  SiteDefinition?: Maybe<SiteDefinitionOutput>;
};
// ...
export type StartQueryVariables = Exact<{
  relativePath?: InputMaybe<Scalars['String']>;
  locales?: InputMaybe<Locales>;
  language?: InputMaybe<Scalars['String']>;
  stageName?: InputMaybe<Scalars['String']>;
  artistGenre?: InputMaybe<Scalars['String']>;
  contentId?: InputMaybe<Scalars['Int']>;
  workId?: InputMaybe<Scalars['Int']>;
  statusEqual?: InputMaybe<Scalars['String']>;
  isCommonDraft?: InputMaybe<Scalars['Boolean']>;
}>;
// ...
export const ArtistDetailsPageFragmentDoc = gql`
    fragment ArtistDetailsPage on ArtistDetailsPage {
  PerformanceStartTime
  PerformanceEndTime
  StageName
  ArtistName
  ArtistPhoto
  ArtistGenre
  ArtistDescription
  ArtistIsHeadliner
  Name
  RelativePath
  ParentLink {
    Url
  }
}
    `;
// ...
export const LandingPageFragmentDoc = gql`
    fragment LandingPage on LandingPage {
  Title
  Subtitle
  BuyTicketBlock {
    ...LandingPageBlockData
  }
  HeroImage
  ArtistsLink {
    Expanded {
      Name
      RelativePath
    }
  }
  MainContentArea {
    ContentLink {
      Expanded {
        ...ItemsInContentArea
      }
    }
  }
  FooterContentArea {
    ContentLink {
      Expanded {
        ...ItemsInContentArea
      }
    }
  }
}
    ${LandingPageBlockDataFragmentDoc}
${ItemsInContentAreaFragmentDoc}`;
export const StartDocument = gql`
    query Start($relativePath: String, $locales: Locales, $language: String, $stageName: String, $artistGenre: String, $contentId: Int, $workId: Int, $statusEqual: String, $isCommonDraft: Boolean) {
  Content(
    locale: [$locales]
    where: {ContentLink: {WorkId: {eq: $workId}, Id: {eq: $contentId}}, RelativePath: {eq: $relativePath}, Language: {Name: {eq: $language}}, Status: {eq: $statusEqual}, IsCommonDraft: {eq: $isCommonDraft}}
    orderBy: {Saved: DESC}
    limit: 1
  ) {
    items {
      Name
      ParentLink {
        Url
      }
      Url
      __typename
      RelativePath
      ... on LandingPage {
        ...LandingPage
        _children {
          ArtistContainerPage {
            items {
              Name
              RelativePath
              headlines: _children {
                ArtistDetailsPage(
                  where: {ArtistIsHeadliner: {eq: true}}
                  orderBy: {PerformanceStartTime: ASC, Name: ASC}
                ) {
                  items {
                    ...ArtistDetailsPage
                  }
                }
              }
            }
          }
        }
      }
      ... on ArtistContainerPage {
        Name
        RelativePath
        artists: _children {
          ArtistDetailsPage(
            where: {StageName: {eq: $stageName}, ArtistGenre: {eq: $artistGenre}, Status: {eq: "Published"}}
            orderBy: {Name: ASC, ArtistIsHeadliner: ASC, PerformanceStartTime: ASC, StageName: ASC}
            limit: 100
          ) {
            items {
              ...ArtistDetailsPage
            }
            facets {
              ArtistGenre(orderType: VALUE, orderBy: ASC, limit: 10) {
                name
                count
              }
              StageName(orderType: VALUE, orderBy: ASC, limit: 10) {
                name
                count
              }
            }
          }
        }
      }
      ... on ArtistDetailsPage {
        ...ArtistDetailsPage
      }
      ... on ContentBlock {
        Title
        Image
        ImageAlignment
        Content
      }
    }
  }
}
    ${LandingPageFragmentDoc}
${ArtistDetailsPageFragmentDoc}`;